upm-subtree-split@v1.1.1
- convert to TypeScript
