+++ package.json
@@ -1,23 +1,26 @@
 {
   "name": "upm-subtree-split",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "GitHub action to perform subtree split to upm branch.",
-  "main": "src/index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "ncc build src/index.js -o dist --source-map --license licenses.txt"
-  },
+  "author": "RageAgainstThePixel",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/RageAgainstThePixel/upm-subtree-split.git"
   },
-  "keywords": [],
-  "author": "Stephen Hodgson",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/RageAgainstThePixel/upm-subtree-split/issues"
   },
-  "homepage": "https://github.com/RageAgainstThePixel/upm-subtree-split#readme",
+  "homepage": "https://github.com/RageAgainstThePixel/upm-subtree-split",
+  "keywords": [
+    "upm",
+    "subtree",
+    "split",
+    "unity",
+    "github",
+    "action"
+  ],
+  "main": "dist/index.js",
   "dependencies": {
     "@actions/core": "^1.10.1",
     "@actions/exec": "^1.1.1",
@@ -25,6 +28,14 @@
     "@actions/io": "^1.1.3"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.34.0"
+    "@types/node": "^22.2.0",
+    "@vercel/ncc": "^0.34.0",
+    "shx": "^0.3.4",
+    "typescript": "^5.5.4"
+  },
+  "scripts": {
+    "build": "npm run clean && npm run bundle",
+    "bundle": "ncc build src/index.ts -o dist --source-map --license licenses.txt",
+    "clean": "npm install && shx rm -rf dist/ out/ node_modules/ && npm ci"
   }
 }
