+++ src/index.ts
@@ -1,0 +1,28 @@
+
+import core = require('@actions/core');
+import exec = require('@actions/exec');
+import glob = require('@actions/glob');
+
+const WORKSPACE = process.env.GITHUB_WORKSPACE;
+
+const main = async () => {
+    try {
+        const packageRootInput = core.getInput(`package-root`) || `${WORKSPACE}/**/Packages/com.*`;
+        let packagePath = undefined;
+        const globber = await glob.create(packageRootInput, { followSymbolicLinks: false });
+        const files = await globber.glob();
+        if (files.length === 0) {
+            throw Error('No package found.');
+        }
+        if (files.length > 1) {
+            throw Error('Multiple packages found!');
+        }
+        packagePath = files[0].replace(`${WORKSPACE}/`, '');
+        await exec.exec('git', ['subtree', 'split', '--prefix', packagePath, '-b', 'upm']);
+        await exec.exec('git', ['push', '-u', 'origin', 'upm', '--force']);
+    } catch (error) {
+        core.setFailed(error.message);
+    }
+}
+
+main();
